Add images legacy processo
