+++ filesave/src/main/java/com/chicco/filesave/dataaccess/DownloadsSaveLegacyProcessor.kt
@@ -4,9 +4,6 @@ import android.os.Environment
 import android.os.Environment.getExternalStoragePublicDirectory
 import com.chicco.filesave.domain.FileContent
 import java.io.File
-import java.io.FileOutputStream
-import java.io.IOException
-import java.io.InputStream
 
 class DownloadsSaveLegacyProcessor : FileSaveProcessor() {
 
@@ -14,31 +11,7 @@ class DownloadsSaveLegacyProcessor : FileSaveProcessor() {
         val directory =
             getExternalStoragePublicDirectory(Environment.DIRECTORY_DOWNLOADS).toString() +
                     file.subfolderName?.let { "/$it" }.orEmpty()
-        val fileName = file.fileNameWithSuffix
 
-        getFile(fileName, file.data, File(directory))
-    }
-
-    private fun getFile(
-        fileName: String,
-        stream: InputStream,
-        attachmentPath: File
-    ): File {
-        val uniqueFileName = FileNameLegacyResolver.getUniqueFileName(attachmentPath, fileName)
-        val savedFile = File(attachmentPath, uniqueFileName)
-        try {
-            savedFile.parentFile?.mkdirs()
-
-            FileOutputStream(savedFile).use { outputStream ->
-                val readBytes = stream.buffered().readBytes()
-                outputStream.write(readBytes)
-                outputStream.flush()
-            }
-        } catch (e: IOException) {
-            savedFile.delete()
-            e.printStackTrace()
-            throw e
-        }
-        return savedFile
+        file.data.saveToFile(file.fileNameWithSuffix, File(directory))
     }
 }
