+++ filesave/src/main/java/com/chicco/filesave/dataaccess/FileSaveExtensions.kt
@@ -4,6 +4,9 @@ import android.content.ContentResolver
 import android.content.ContentValues
 import android.net.Uri
 import android.os.ParcelFileDescriptor
+import java.io.File
+import java.io.FileOutputStream
+import java.io.IOException
 import java.io.InputStream
 
 fun ContentResolver.saveFile(
@@ -18,5 +21,29 @@ fun ContentResolver.saveFile(
                     stream.buffered().readBytes()
                 )
             }
+    }
+}
+
+fun InputStream.saveToFile(
+    fileName: String,
+    attachmentPath: File
+): File {
+    use {
+        val uniqueFileName = FileNameLegacyResolver.getUniqueFileName(attachmentPath, fileName)
+        val savedFile = File(attachmentPath, uniqueFileName)
+        try {
+            savedFile.parentFile?.mkdirs()
+
+            FileOutputStream(savedFile).use { outputStream ->
+                val readBytes = buffered().readBytes()
+                outputStream.write(readBytes)
+                outputStream.flush()
+            }
+        } catch (e: IOException) {
+            savedFile.delete()
+            e.printStackTrace()
+            throw e
+        }
+        return savedFile
     }
 }
