+++ filesave/src/main/java/com/chicco/filesave/domain/FileContent.kt
@@ -6,7 +6,7 @@ import java.io.InputStream
 data class FileContent(
     val data: InputStream,
     val fileNameWithoutSuffix: String,
-    val suffix: String?,
+    val suffix: String,
     val mimeType: String?,
     val subfolderName: String?
 ) {
