+++ filesave/src/main/java/com/chicco/filesave/usecase/FileSaveController.kt
@@ -2,10 +2,7 @@ package com.chicco.filesave.usecase
 
 import android.content.Context
 import android.os.Build
-import com.chicco.filesave.dataaccess.DownloadsSaveLegacyProcessor
-import com.chicco.filesave.dataaccess.DownloadsSaveProcessor
-import com.chicco.filesave.dataaccess.FileSaveProcessor
-import com.chicco.filesave.dataaccess.ImageFileSaveProcessor
+import com.chicco.filesave.dataaccess.*
 import com.chicco.filesave.domain.FileContent
 
 class FileSaveController(
@@ -17,8 +14,12 @@ class FileSaveController(
         } else {
             DownloadsSaveLegacyProcessor()
         }
-    private val imagesFileSaveProcessor: ImageFileSaveProcessor =
-        ImageFileSaveProcessor(context.contentResolver)
+    private val imagesFileSaveProcessor: FileSaveProcessor =
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.Q) {
+            ImageFileSaveProcessor(context.contentResolver)
+        } else {
+            ImageFileSaveLegacyProcessor()
+        }
 
     suspend fun savePdfFile(content: FileContent) {
         downloadsProcessor.save(content)
